modidfy

+++ index.js
@@ -60,30 +60,42 @@ async function run() {
     // products curd
     app.get("/products/:id", async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const query = { _id: new ObjectId(id) };
       const result = await productCollection.findOne(query);
-      res.send(result);
+      return res.send(result);
     });
     app.get("/products/categories/:category", async (req, res) => {
       const category = req.params.category;
+      if(!category){
+        return res.send({message:"category Notfound"})
+      }
       const query = { category };
       const result = await productCollection.find(query).toArray();
-      res.send(result);
+      return res.send(result);
     });
     app.patch("/products/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const product = req.body;
       const result = await productCollection.updateOne(filter, {
         $set: product,
       });
-      res.send(result);
+      return res.send(result);
     });
     app.delete("/products/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const result = await productCollection.deleteOne(filter);
-      res.send(result);
+      return res.send(result);
     });
     app.get("/products", async (req, res) => {
       const query = {};
@@ -92,49 +104,67 @@ async function run() {
     });
     app.post("/products", verifyToken, async (req, res) => {
       const product = req.body;
+      if(!product){
+        return res.send({message:"No data for insert"})
+      }
       const result = await productCollection.insertOne(product);
-      res.send(result);
+      return res.send(result);
     });
 
     // Blogs curd
     app.get("/blogs/:id", async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const query = { _id: new ObjectId(id) };
       const result = await blogCollection.findOne(query);
-      res.send(result);
+      return res.send(result);
     });
     app.patch("/blogs/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const blog = req.body;
       const result = await blogCollection.updateOne(filter, {
         $set: blog,
       });
-      res.send(result);
+      return res.send(result);
     });
     app.delete("/blogs/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const result = await blogCollection.deleteOne(filter);
-      res.send(result);
+     return res.send(result);
     });
     app.get("/blogs", async (req, res) => {
       const query = {};
       const result = await blogCollection.find(query).toArray();
-      res.send(result);
+      return res.send(result);
     });
     app.post("/blogs", verifyToken, async (req, res) => {
       const blog = req.body;
+      if(!blog){
+        return res.send({message:"No data for insert"})
+      }
       const result = await blogCollection.insertOne(blog);
-      res.send(result);
+      return res.send(result);
     });
 
     // reviews curd
     app.get("/reviews/:id", async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const query = { _id: new ObjectId(id) };
       const result = await reviewCollection.findOne(query);
-      res.send(result);
+      return res.send(result);
     });
     app.get("/reviews/me/:email", async (req, res) => {
       const email = req.params.email;
@@ -144,18 +174,24 @@ async function run() {
     });
     app.patch("/reviews/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const review = req.body;
       const result = await reviewCollection.updateOne(filter, {
         $set: review,
       });
-      res.send(result);
+      return res.send(result);
     });
     app.delete("/reviews/:id", verifyToken, async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const result = await reviewCollection.deleteOne(filter);
-      res.send(result);
+       return res.send(result);
     });
     app.get("/reviews", async (req, res) => {
       const query = {};
@@ -164,54 +200,72 @@ async function run() {
     });
     app.post("/reviews", verifyToken, async (req, res) => {
       const review = req.body;
+      if(!review){
+        return res.send({message:"No data for insert"})
+      }
       const result = await reviewCollection.insertOne(review);
-      res.send(result);
+      return res.send(result);
     });
 
     // users curd
     app.get("/users/get/:id", async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const query = { _id: new ObjectId(id) };
       const result = await userCollection.findOne(query);
-      res.send(result);
+      return res.send(result);
     });
     app.get("/users/:email", async (req, res) => {
       const email = req.params.email;
+      if(!email){
+        return res.send({message:"Email not fond"})
+      }
       const query = { email };
       const result = await userCollection.findOne(query);
-      res.send(result);
+     return res.send(result);
     });
     app.patch("/users/:id", async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const newDoc = req.body;
       const result = await userCollection.updateOne(filter, {
         $set: newDoc,
       });
-      res.send(result);
+     return res.send(result);
     });
     app.delete("/users/:id", async (req, res) => {
       const id = req.params.id;
+      if(!id){
+        return res.send({message:"Id invalid"})
+      }
       const filter = { _id: new ObjectId(id) };
       const result = await userCollection.deleteOne(filter);
-      res.send(result);
+      return res.send(result);
     });
 
     //verify admin
-    app.get("/users", async (req, res) => {
+    app.get("/users",verifyToken, async (req, res) => {
       const query = {};
       const result = await userCollection.find(query).toArray();
       res.send(result);
     });
     app.post("/users", async (req, res) => {
       const user = req.body;
+      if(!user){
+        return res.send({message:"Id invalid"})
+      }
       const query = { email: user.email };
       const exist = await userCollection.findOne(query);
       if (exist) {
         return res.send({ message: "user already exist" });
       }
       const result = await userCollection.insertOne({ ...user, role: "user" });
-      res.send(result);
+      return res.send(result);
     });
   } finally {
     // Ensures that the client will close when you finish/error
